separated internal scales from exports to resolve circular dependencies
