+++ packages/rating-scales/src/discrete/DiscreteScale.js
