+++ packages/rating-scales/src/ellipse/EllipseScale.js
