+++ packages/rating-scales/src/index.js
@@ -1,4 +1,4 @@
-export * from "./discrete/Scale";
-export * from "./ellipse/Scale";
-export * from "./vas/Scale";
-export * from "./vas/MultiScale";
+export * from "./discrete/DiscreteScale";
+export * from "./ellipse/EllipseScale";
+export * from "./vas/VisualAnalogScale";
+export * from "./vas/MultiVisualAnalogScale";
