+++ packages/rating-scales/src/vas/Scale.js
@@ -1,7 +1,5 @@
 import { useState, useCallback } from "react";
 import UnitValue from "unit-value/lib/unit-value";
-import Frame from "../core/Frame";
-import Question from "../core/Question";
 import {
   FlexContainer,
   ScaleBar,
@@ -14,8 +12,6 @@ import {
   getValueForRelativeX,
   getXPosForValue,
 } from "../core/services/bar-coords";
-import { Confidence } from "./Confidence";
-import { valueIds } from "./constants";
 import { DragMarker } from "./DragMarker";
 
 export const Scale = ({
@@ -112,79 +108,3 @@ export const Scale = ({
     </ScaleBar>
   );
 };
-
-/**
- * State Hook for using a VisualAnalogScale
- * @param {*} initialValue
- * @param {*} initialConfidence
- * @returns
- */
-export const useVisualAnalogScale = (initialValues = {}) => {
-  const [values, setValues] = useState(initialValues);
-
-  const onChange = (id, v, newValues) => setValues({ ...newValues, [id]: v });
-
-  return {
-    props: { values },
-    handlers: { onChange },
-    setValues,
-  };
-};
-
-const VisualAnalogScale = ({
-  frameHeight = "300px",
-  questionOptions = {},
-  question,
-  barOptions,
-  labels,
-  labelOptions,
-  scaleMarkerOptions,
-  dragMarkerOptions,
-  values = {},
-  onChange = () => {},
-  useConfidenceInput,
-  confidenceText = "How confident are you?",
-  confidenceTextOptions = { topMargin: "0%", xAlign: "center" },
-}) => {
-  const handleScaleChange = (value) => {
-    onChange(valueIds.scale, value, { ...values, [valueIds.scale]: value });
-  };
-
-  const handleConfidenceChange = (value) => {
-    value = Math.min(Math.max(value, 0), 100);
-    onChange(valueIds.confidence, value, {
-      ...values,
-      [valueIds.confidence]: value,
-    });
-  };
-
-  return (
-    <>
-      <Frame frameHeight={frameHeight}>
-        <Question {...questionOptions}>{question}</Question>
-        <Scale
-          barOptions={barOptions}
-          labels={labels}
-          labelOptions={labelOptions}
-          scaleMarkerOptions={scaleMarkerOptions}
-          dragMarkerOptions={dragMarkerOptions}
-          onChange={handleScaleChange}
-          value={values.scale}
-        />
-      </Frame>
-      {useConfidenceInput && (
-        <Confidence
-          confidenceText={confidenceText}
-          confidenceTextOptions={confidenceTextOptions}
-          isDisabled={values.scale == null}
-          topMargin={frameHeight}
-          onChange={handleConfidenceChange}
-          value={values.confidence}
-          style={useConfidenceInput}
-        />
-      )}
-    </>
-  );
-};
-
-export { VisualAnalogScale };
